remark,grade,schedule

+++ src/main/java/com/byt/freeEdu/service/GradeService.java
@@ -1,44 +1,75 @@
 package com.byt.freeEdu.service;
 
+import com.byt.freeEdu.mapper.GradeMapper;
+import com.byt.freeEdu.model.DTO.GradeDto;
 import com.byt.freeEdu.model.Grade;
+import com.byt.freeEdu.model.enums.SubjectEnum;
+import com.byt.freeEdu.model.users.Student;
+import com.byt.freeEdu.model.users.Teacher;
 import com.byt.freeEdu.repository.GradeRepository;
+import com.byt.freeEdu.repository.TeacherRepository;
+import com.byt.freeEdu.service.users.StudentService;
+import com.byt.freeEdu.service.users.TeacherService;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class GradeService {
     private final GradeRepository gradeRepository;
+    private final TeacherRepository teacherRepository;
+    private final GradeMapper gradeMapper;
+    private final StudentService studentService;
+    private final TeacherService teacherService;
 
-    public GradeService(GradeRepository gradeRepository) {
+    public GradeService(GradeRepository gradeRepository, TeacherRepository teacherRepository, GradeMapper gradeMapper, StudentService studentService, TeacherService teacherService) {
         this.gradeRepository = gradeRepository;
+        this.teacherRepository = teacherRepository;
+        this.gradeMapper = gradeMapper;
+        this.studentService = studentService;
+        this.teacherService = teacherService;
     }
 
-    public Grade saveGrade(Grade grade) {
-        return gradeRepository.save(grade);
+    public Boolean saveGrade(GradeDto grade) {
+        Grade entity = gradeMapper.toEntity(grade, studentService, teacherService);
+        entity.setSubject(SubjectEnum.valueOf(grade.getSubject()));
+        entity.setGradeDate(LocalDate.now());
+        gradeRepository.save(entity);
+        return true;
     }
 
     public Grade getGradeById(int id) {
         return gradeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Grade not found with ID: " + id));
+    }
+    public List<GradeDto> getGradesByTeacherId(int teacherId) {
+        return gradeRepository.getGradeByTeacher(teacherRepository.findById(teacherId).get())
+                .stream()
+                .map(gradeMapper::toDto)
+                .collect(Collectors.toList());
     }
 
     public List<Grade> getAllGrades() {
         return gradeRepository.findAll();
     }
 
-    public Grade updateGrade(int id, Grade updatedGrade) {
-        Grade existingGrade = gradeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Grade not found with ID: " + id));
-        existingGrade.setSubject(updatedGrade.getSubject());
+    public Boolean updateGrade(int id, GradeDto updatedGrade) {
+        Grade existingGrade = gradeRepository.findById(id).get();
         existingGrade.setValue(updatedGrade.getValue());
-        existingGrade.setStudent(updatedGrade.getStudent());
-        existingGrade.setTeacher(updatedGrade.getTeacher());
-        existingGrade.setGradeDate(updatedGrade.getGradeDate());
+        existingGrade.setGradeDate(LocalDate.now());
 
-        return gradeRepository.save(existingGrade);
+        gradeRepository.save(existingGrade);
+        return true;
     }
 
-    public void deleteGrade(int id) {
-        gradeRepository.deleteById(id);
+    public boolean deleteGrade(int id) {
+        if (gradeRepository.existsById(id)) {
+            gradeRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Ocena o podanym ID nie istnieje.");
+        }
+        return true;
     }
 }
